% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_estimate.R
\docType{data}
\name{anchor_estimate}
\alias{anchor_estimate}
\title{Two Independent Groups - Anchor Adjust Effect}
\format{A data frame with 90 rows and 8 variables:
\describe{
 \item{sessionid}{Respondent identifier}
 \item{location}{Lab that ran the study.  Equal to "Ithica" for all participants in this dataset.}
 \item{babies_anchor}{Babies Anchor: "lowanchor" (100) or "highanchor" (50,000)}
 \item{babies_est}{Participants' estimate of number of  babies born per day}
 \item{chicago_anchor}{Chicago Anchor: "low anchor" (200,000) or "high anchor" (5,000,000)}
 \item{chicago_est}{Participants' estimate of the population of Chicago}
 \item{everest_anchor}{Everest Anchor: "low anchor" (2,000 feet) or "high anchor" (45,000 feet)}
 \item{everest_est}{Participants' estimate of height of Mt Everest (in feet)}
}}
\source{
The data is available online at
 \url{https://osf.io/wx7ck/} and is from one replication site (Ithaca) of the following study:

Klein, R. A., Ratliff, K. A., Vianello, M., Adams ., R. B., Bahnik, S., Bernstein, M. J., ... & Nosek, B. A. (2014).
Investigating Variation in Replicability. \emph{Social Psychology, 45}, 142-152. \url{http://doi.org/10.1027/1864-9335/a000178}

 The original study exploring this effect is: Jacowitz, K. E., & Kahneman, D.
 (1995). Measures of Anchoring in Estimation Tasks. \emph{Personality and
 Social Psychology Bulletin, 21}, 1161-1166.
 \url{http://doi.org/10.1177/01461672952111004}
}
\usage{
anchor_estimate
}
\description{
An example of data from a study with a two independent groups design used in Chapter 7 of the book
\emph{Introduction to the New Statistics}.
}
\details{
Psychological research suggests that how a question is worded can influence people's judgments.
This is known as the 'anchor-adjustment' effect.  A team of psychologists conducted a multi-site study
attempting to replicate this phenomenon. Participants were asked to estimate three different quantities (number of
babies born in the U.S. each day; population of Chicago; height of Mounter
Everest).  For each question, though, participants were either given a low or
high numerical anchor (see below).  For example, they were told either that
the number of babies born in the U.S. was more than 200,000 (low anchor) or
less than 5,000,000 (high anchor).  The primary research question was: to what extent does
having a low or high anchor in mind influence the estimate made?

The anchors were as follows:

\emph{Babies}

Low: More than 100 babies are born per day in the United States. How many babies do you think are born in the U.S. each day?

High: Less than 50,000 babies are born per day in the United States. How many babies do you think are born in the U.S. each day?

\emph{Population}

Low: The population of Chicago is more than 200,000. What do you think the population of Chicago is?

High: The population of Chicago is less than 5,000,000. What do you think the population of Chicago is?

\emph{Everest}

High: Mount Everest is shorter than 45,500 feet. How tall do you think Mount Everest is?

Low: Mount Everest is taller than 2,000 feet.  How tall do you think Mount Everest is?
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

