% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blame.R
\docType{data}
\name{blame2}
\alias{blame2}
\title{Two Way Design - Blame 2}
\format{A data frame with 432 rows and 4 variables:
\describe{
 \item{id}{Participant ID}
 \item{race}{\emph{black} or \emph{white}}
 \item{ses}{Socio-economic status: \emph{high} or {low}}
 \item{blame}{Rating of blame on 1 to 6 scale}
}}
\source{
You can find the materials and data on the Open Science Framework: \url{https://osf.io/xtkve}
}
\usage{
blame2
}
\description{
An example dataset used in Chapter 15 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
The data here is from a study that was a direct replication of that described in \code{\link{blame1}}.

A new set of online U.S. participants again read about a crime being committed.  Indicators of race and SES were again
varied, and participants rate blame on a scale from 1 to 6.

One key difference with this replication is that the description of the crime was altered to make it clear that
the perpetrator was guilty.  Thus, blame ratings are higher than in the first study, where it was slightly ambiguous
if the person described has committed the crime.

This is a simplified dataset from the study, that has been graciously shared by the authors.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{blame1}}
}
\keyword{datasets}

