% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_well.R
\docType{data}
\name{body_well}
\alias{body_well}
\title{Correlated Data - Body Wellness}
\format{A data frame with 106 rows and 3 variables:
\describe{
 \item{sex}{\emph{female} or \emph{male}}
 \item{bodysat}{Body Satisfaction score}
 \item{wellbeing}{Wellbeing score}
}}
\usage{
body_well
}
\description{
An example of correlated data used in Chapter 11 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
This dataset contains body satisfaction and wellbeing scores for a sample of 106 college students.
The \emph{bodysat} variable is the mean rating of how satisfied a person is
(from 1 = very dissatisfied to 5 = very satisfied) with various aspects of
their body (e.g. satisfaction with one's face, muscle tone, 
weight, etc.). It is a subscale of the Multidimensional Body Self-Relations
Questionnaire (MBSRQ). The \emph{wellbeing} variable is the mean rating of strength
of agreement (from 1 = strongly disagree to 7 = strongly agree) with a number
of statements about a person's feeling of wellbeing.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}
