% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campus_involvement.R
\docType{data}
\name{campus_involvement}
\alias{campus_involvement}
\title{Correlated Data - Campus Involvement}
\format{A data frame with 113 rows and 6 variables:
\describe{
 \item{responseid}{Respondent Identifier}
 \item{gender}{\emph{Female} or \emph{Male}}
 \item{gpa}{Grade Point Average}
 \item{commuter}{Variable indicating whether student lived on campus (\emph{resident}) or off campus (\emph{commuter})}
 \item{swb}{Average score on Diener's Subjective Wellbeing Scale. Scores should range from 1 to 5.}
 \item{campus_involvement}{Average of an 8-item scale of campus involvement. Scores should range between 1 and 5.}
}}
\usage{
campus_involvement
}
\description{
An example of correlated data used in Chapter 11 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
The data were collected as part of an online survey of college students.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

