% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_moral.R
\docType{data}
\name{clean_moral2}
\alias{clean_moral2}
\title{Two Independent Groups - Clean Moral Replication Study}
\format{A data frame with 208 rows and 3 variables:
\describe{
 \item{participant}{Respondent identifier}
 \item{condition}{Type of words unscrambled: neutral priming or clean priming}
 \item{moral_judgements}{Average of moral judgement scale, 6 items, all ranked from 0-9}
}}
\usage{
clean_moral2
}
\description{
An example of data from a study with a two independent groups design used in Chapter 7 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This data comes from a replication of the study described in \code{\link{clean_moral1}}

After publication of that study, Cheung, Donnellan & Johnson (2014) attempted a close replication,
using the same manipulation and the same measure of moral judgement.
}
\section{Source}{

The dataset is available online at:  \url{https://osf.io/apidb}  and is Study 1 from:

Johnson, D. J., Cheung, F., & Donnellan, M. B. (2014). Does cleanliness influence moral judgments?: A direct replication of Schnall, Benton, and Harvey (2008).
\emph{Social Psychology, 45}, 209-215. \url{doi:10.1027/1864-9335/a000186}

Note that this a cleaned up version of the online dataset - all participants excluded for various reasons (e.g. experimenter error) have been deleted.
}

\section{References}{

Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{clean_moral1}}
}
\keyword{datasets}

