% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_moral.R
\docType{data}
\name{clean_moral_schall}
\alias{clean_moral_schall}
\title{Two Independent Groups - Clean Moral Original Study}
\format{A data frame with 40 rows and 3 variables:
\describe{
 \item{participant}{Respondent identifier}
 \item{condition}{Type of words unscrambled: \emph{neutral priming} or \emph{clean priming}}
 \item{moral_judgements}{Average of moral judgement scale, 6 items, all ranked from 0-9}
}}
\source{
The dataset is available online \url{https://osf.io/4cs3k} and is Study 1 from:

Schnall, S., Benton, J., & Harvey, S. (2008). With a Clean Conscience: Cleanliness Reduces the Severity of Moral Judgments.
\emph{Psychological Science, 19}, 1219-1222. \url{http://doi.org/10.1111/j.1467-9280.2008.02227.x}
}
\usage{
clean_moral_schall
}
\description{
An example of data from a study with a two independent groups design used in Chapter 7 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
Some researchers believe that moral judgements are based not only on rational considerations but on one's current emotional state.
To what extent can recent emotional experiences influence moral judgements?  Schnall et al. (2008) examined this question by manipulating
feelings of cleanliness and purity and then observing if this changes how harshly participants judged the morality of others.
Participants completed a word scramble task with either neutral words (neutral priming) or words related to
cleanliness (cleanliness prime).  All students then completed a set of moral judgements about controversial scenarios
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{clean_moral_johnson}, which is a replication of the Schnall et al study}
}
\keyword{datasets}

