% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensd_rm.R
\name{cohensd_rm}
\alias{cohensd_rm}
\title{Cohen's d standardized mean change effect size for repeated measures designs}
\usage{
cohensd_rm(x, y, ci = 95, unbiased = FALSE)
}
\arguments{
\item{x}{Numeric vector of observations at time 1 (e.g., pre-test)}

\item{y}{Numeric vector of observations at time 2 (e.g., post-test)}

\item{ci}{Confidence level.  Default is 95 (for a 95 percent CI).}

\item{unbiased}{Logical. If TRUE, the estimated effect size is corrected for 
small-sample bias.  Default is FALSE.}
}
\value{
A numeric vector of length three comprising the estimated effect size 
 (est), lower limit of the confidence interval (ll), and upper limit of the 
 confidence interval (ul).
}
\description{
Computes Cohen's d effect size for repeated measures designs (paired samples),
using the standardizer recommended by Cumming (2012).  In other words, the 
standardizer is the average of the pre and post-treatment standard deviations,
rather than the standard deviation of the change scores.  An approximate 
noncentral-t confidence interval is computed using the method proposed by 
Algina & Keselman (2003), Equations 7 to 9.  The effect size estimate can be 
corrected for sample sample bias (see Cumming, 2012, p.294) by setting the 
\emph{Unbiased} argument to \emph{TRUE}.
}
\examples{
\dontrun{
   cohensd_rm(x = thomason1$pre, y = thomason1$post)
}
}
\references{
Algina, J. A., & Keselman, H. J. (2003). Approximate Confidence
Intervals for Effect Sizes. \emph{Educational and Psychological Measurement,
63}, 537-553.

Cumming, G. (2012). \emph{Understanding The New Statistics:
Effect Sizes, Confidence Intervals, and Meta-Analysis}. Routledge; New York.
}

