% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_surveys.R
\docType{data}
\name{college_survey1}
\alias{college_survey1}
\title{Describing Data Example - College Student Attitude Survey 1}
\format{A data frame with 243 rows and 20 variables:
\describe{
 \item{id}{Participant identifier}
 \item{gender}{Sex: Female or Male}
 \item{age}{Age in years}
 \item{school_year}{5 level school year}
 \item{transfer}{Transfer status: \emph{No} if student did not transfer to current school, \emph{Yes} if student transferred to current school}
 \item{student_athlete}{3-level student athlete code}
 \item{student_athlete_code}{2-level student athlete code}
 \item{wealth_sr}{Self-rating of perceived wealth, rated on scale from 1 (well below average) to 5 (well above average).}
 \item{gpa}{Self reported GPA on a 0-4point scale.}
 \item{act}{Self-reported ACT score.}
 \item{subjective_well_being}{verage of 5-item satisfaction with life scale.}
 \item{positive_affect}{Average of 10-item scale of frequency of experiencing positive emotions over the past week.}
 \item{negative_affect}{Average of 10-item scale of frequency of experiencing negative emotions over the past week.}
 \item{relationship_confidence}{Average of a 7-item scale of confidence in having romantic relationships, rated on a scale from 1 (not at all characteristic of me) to 5 (very characteristic of me).}
 \item{exercise}{Exercise score on the Godin Leisure-Time Exercise Quetionnaire.  This asks participants to rate how often during a regular week they engage in strenuous, moderate, or light exercise.}
 \item{academic_motivation_intrinsic}{Average score on 6-item measure of intrinsic reasons for going to college.}
 \item{academic_motivation_extrinsic}{Average score on 7-item measure of extinsic reasons for going to college.}
 \item{academic_motivation_amotivation}{Average score on 2-item measure of feeling unmotivated about college}
 \item{intelligence_value}{Measure of degree to which participant values intelligence, rated on a scale from 1 (strongly agree) to 5 (strongly agree).}
 \item{raven_score}{Percentage correct out of 8 items on the Raven Progressive Matrix Scale, a scale of logical thinking/IQ}.
}}
\usage{
college_survey1
}
\description{
An example of data from a college survey used in Chapters 3 and 5 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This data comes from a survey of college students on a range of psychological variables.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

