% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dana.R
\docType{data}
\name{dana}
\alias{dana}
\title{Impact of Outliers - Target Contact}
\format{A data frame with 38 rows and 2 variables:
\describe{
 \item{group}{Group: group one or group two (independent variable)}
 \item{contact}{Contact time in seconds (dependent variable)} }}
\source{
Wilcox, R. R. (2009). \emph{Basic statistics: Understanding
 conventional methods and modern insights}. New York: Oxford University
 Press.
}
\usage{
dana
}
\description{
An example of independent groups data with outliers used in Chapters 16 of the
book \emph{Introduction to the New Statistics}.
}
\details{
The data come from a study by Dana (1990, cited in Wilcox 2009) in which
individuals tried to keep an apparatus in contact with a target. There was
one experimental condition with two levels (group one and group two). The
dependent variable is the number of seconds participants could keep in contact
with the target.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

