% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{demo_fun}
\alias{demo_fun}
\title{Demo function title here}
\usage{
demo_fun(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A second numeric vector.}
}
\value{
Return demo value here
}
\description{
\code{demo_fun} The second paragraph is the description. is used to sum two
numbers and I want to make it well longer than 80 characters ok.
}
\section{MyTestCustomSection}{

 here is my test custom section does it work
}
\examples{
demo_fun(1, 10)
demo_fun(1000, 3)

\dontrun{
demo_fun("a", 3)
}
}
\seealso{
\code{\link{hello}} for more info
}

