% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escir.R
\docType{package}
\name{escir}
\alias{escir}
\alias{escir-package}
\title{Effect Size and Confidence Intervals in R.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Demo functions}{

The foo functions ...
}

