% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_scores.R
\docType{data}
\name{exam_scores}
\alias{exam_scores}
\title{Correlated Data - Exam Scores}
\format{A data frame with 9 rows and 3 variables:
\describe{
 \item{id}{Respondent ID}
 \item{exam1}{Exam score - first occasion}
 \item{exam_final}{Exam score - last occasion}
}}
\usage{
exam_scores
}
\description{
An example of correlated data used in Chapter 11 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
This is a set of exam scores from students in an introductory psychology course.
Student IDs are fake, but the data is real.  Scores are expressed as a percentage,
where 0 = no correct and 100 = all correct.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

