% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habituation.R
\docType{data}
\name{habituation}
\alias{habituation}
\title{Two Dependent Groups - Habituation}
\format{A data frame with 14 rows and 4 variables:
\describe{
 \item{animalid}{ID of the animal}
 \item{condition}{Factor with one level - \emph{trained}}
 \item{base_m}{Duration (in seconds) of tail-elicited siphon withdrawal reflex, average of 3 measures, taken prior to start of any training}
 \item{lt3_m}{Duration (in seconds) of tail-elicited siphon withdrawal reflex, average of 3 measures, taken 24h after start of third round of training}
}}
\source{
This data is availble online at \url{https://osf.io/6ew4i} and was reported in Figure 4 of:

Holmes, G., Herdegen, S., Schuon, J., Cyriac, A., Lass, J., Conte, C., ..., Calin-Jageman, R. J. (2014).
Transcriptional analysis of a whole-body form of long-term habituation in Aplysia californica.
\emph{Learning & Memory, 22}, 11-23. \url{http://doi.org/10.1101/lm.036970.114}
}
\usage{
habituation
}
\description{
An example of data from a study with a two dependent groups design used in Chapter 8 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
To what extent does the pattern of training influence learning?  To find out, sea slugs were given
3 different types of experience.  \emph{Control animals} just sat in their aquarium.  \emph{Trained slugs} received
3 days of habituation training - this was basically just brushing the sea slug for 15 minutes, then
taking a 15 minute break, then repeating. This type of training causes the sea slug to learn to
ignore the brushing.  The last group of slugs received \emph{massed training}, meaning that they got the
same brushing as the trained slugs but all packed into a single day with no breaks whatsoever.

To measure learning, reflexes were times before (base_M) and 24 hours after the training (lt3_m)

This is a simplified data set that contains data for only the trained animals.  In addition,
only the pre-test and post-test measures are included.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

