% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_prices.R
\docType{data}
\name{home_prices}
\alias{home_prices}
\title{Regression Example - Home Prices}
\format{A data frame with 300 rows and 7 variables:
\describe{
 \item{mls}{Multiple listing service number for the house (unique ID)}
 \item{location}{City/town where the house is located (factor with 27 levels)}
 \item{price}{Most recent listing price of the house (in dollars)}
 \item{bedrooms}{Number of bedrooms}
 \item{bathrooms}{Number of bathrooms}
 \item{size}{Size of the house in square feet}
 \item{status}{Type of sale - \emph{Foreclosure} or\emph{Short Sale}}
}}
\source{
This is a subset of the first 300 rows of the full dataset, which can be found at
\url{https://wiki.csc.calpoly.edu/datasets/wiki/Houses}
}
\usage{
home_prices
}
\description{
An example dataset used in Chapter 12 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This dataset contains a collection of real estate listings in San Luis Obispo county in California, United States.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

