% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_prices_holdout.R
\docType{data}
\name{home_prices_holdout}
\alias{home_prices_holdout}
\title{Regression Example - Home Prices Holdout}
\format{A data frame with 10 rows and 3 variables:
\describe{
 \item{new_case}{Identifier}
 \item{size}{Size of the house in square feet}
 \item{asking_price}{Asking price in dollars}
}}
\source{
This is a subset of data that can be found at
\url{https://wiki.csc.calpoly.edu/datasets/wiki/Houses}
}
\usage{
home_prices_holdout
}
\description{
An example dataset used in Chapter 12 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This is a holdout sample of an additional 10 cases from the Houses dataset. See
\code{\link{home_prices}} for more info.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{home_prices}}
}
\keyword{datasets}

