% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_boost.R
\docType{data}
\name{iq_boost}
\alias{iq_boost}
\title{One Way Design - IQ Boosting Drugs}
\format{A data frame with 60 rows and 2 variables:
\describe{
 \item{drug}{Factor with 6 levels}
 \item{iq}{IQ score}
}}
\usage{
iq_boost
}
\description{
An example dataset used in Chapter 13 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This is a fake dataset created for the purposes of teaching / illustration.

A company is trying to develop a smart drug that can boost IQ.  Because the herbal remedy gingko biloba has been reputed to enhance cognition,
they isolate 5 different compounds from gingko leaves (Drugs A, B, C, D, and E).  To see if any of these have promise,
60 volunteers are recruited and randomly assigned to receive either placebo, Drug A, Drug B, Drug C, Drug D, or Drug E
(let's assume all the compounds are known to be completely safe).  After drug administration, all the participants take an IQ test.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

