% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_genes.R
\docType{data}
\name{learning_genes}
\alias{learning_genes}
\title{Two Dependent Groups - Learning Genes}
\format{A data frame with 12 rows and 7 variables:
\describe{
 \item{animal}{ID of the animal}
 \item{tissue}{\emph{SW} indicates gene measured in sensory neurons that help detect touch to the skin}
 \item{trained_side}{Side on which animal received sensitization training (Left or Right)}
 \item{egr_untrained}{Amount of RNA for the Egr Gene (early growth response) on the trained side of the nervous system}
 \item{egr_trained}{Amount of RNA for the Egr Gene on the untrained side of the nervous system}
 \item{creb_untrained}{Amount of RNA for the CREB gene (Cre element binding protein) on the trained side of the nervous system}
 \item{creb_trained}{Amount of RNA for the CREB gene on the untrained side of the nervous system}
}}
\source{
This data is availble online at \url{https://osf.io/ts9ea/} and was reported in Figure 2 of:

Herdegen, S., Conte, C., Kamal, S., Calin-Jageman, R. J., & Calin-Jageman, I. E. (2014).
Immediate and Persistent Transcriptional Correlates of Long-Term Sensitization Training at Different CNS Loci in Aplysia californica.
\emph{PloS One, 9}, e114481. \url{http://doi.org/10.1371/journal.pone.0114481}
}
\usage{
learning_genes
}
\description{
An example of data from a study with a two dependent groups design used in Chapter 8 of the book
\emph{Introduction to the New Statistics}.
}
\details{
To what extent does storing a memory modify gene expression in the nervous system?  To find out,
animals received sensitization training one one side of the body (left or right).  This produced
a long-term memory only on the side of training.  1 day after training, animals were dissected and
sensory neurons from the left and right side of the body were extracted.  Each sample was then
tested using qPCR to measure the amount of RNA for the Egr or CREB genes.  To what extent does
training persistently alter the amount of Egr RNA?  To what extent does training persistently alter
the amount of CREB RNA?
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

