% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_files.R
\docType{data}
\name{ma_anchor_adjust}
\alias{ma_anchor_adjust}
\title{Meta-Analysis Example - Anchor Estimate}
\format{A data frame with 30 rows and 9 variables: 
\describe{ 
 \item{location}{Name of the lab}
 \item{m_lowanchor}{Mean for those given the low anchor}
 \item{sd_lowanchor}{Standard deviation for those given the low anchor}
 \item{n_lowanchor}{Sample size for those given the low anchor}
 \item{m_highanchor}{Mean for those given the high anchor}
 \item{sd_highanchor}{Standard Deviation for those given the high anchor}
 \item{n_highanchor}{Sample size for those given the high anchor}
 \item{subset}{Factor indicating whether the study was conducted in the USA or not}
 \item{country}{Factor with 12 levels indicating the country where the study was conducted}
}}
\source{
This is data is available online at \url{https://osf.io/wx7ck} from this study:
Klein, R. A., Ratliff, K. A., Vianello, M., Adams, R. B., Bahník, Š., Bernstein, M. J.,
... Nosek, B. A. (2014). Investigating Variation in Replicability. 
\emph{Social Psychology, 45}(3), 142–152. \url{http://doi.org/10.1027/1864-9335/a000178}

The original study exploring this effect is:
Jacowitz, K. E., & Kahneman, D. (1995). Measures of Anchoring in Estimation Tasks. 
\emph{Personality and Social Psychology Bulletin, 21}(11), 1161–1166. 
\url{http://doi.org/10.1177/01461672952111004}
}
\usage{
ma_anchor_adjust
}
\description{
An example dataset used in Chapter 9 of the book \emph{Introduction to the New
Statistics}.
}
\details{
To what extent does the wording of a question influence one's judgement?  This
data investigates a specific type of wording influence: a numerical anchor. 
Participants were asked to estimate three different quantities (number of
babies born in the U.S. each day; population of Chicago; height of Mounter
Everest).  For each question, though, participants were either given a low or
high numerical anchor.  For example, they were told either that
the number of babies born in the U.S. was more than 200,000 (low anchor) or
less than 5,000,000 (high anchor).  The question is: to what extent does
having a low or high anchor in mind influence the estimate made?

This dataset provides summaries for  30 of the 36 different labs that tried to
replicate this classic effect.  This is the data for the estimated number of
babies born/day in the U.S.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017). \emph{Introduction to the
New Statistics}. New York; Routledge.
}
\keyword{datasets}

