% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_files.R
\docType{data}
\name{ma_flag_priming}
\alias{ma_flag_priming}
\title{Meta-Analysis Example - Replication of Flag Priming and Conservatism}
\format{A data frame with 25 rows and 7 variables: 
\describe{ 
 \item{location}{Name of the lab that ran the study}
 \item{m_flagprimed}{Mean conservatism score for flag-primed group. Score can range from 1 to 7,low scores indicate liberal attitudes, high scores indicate conservative attitudes.}
 \item{sd_flagprimed}{Standard deviation for those given the flag-primed group}
 \item{n_flagprimed}{Sample size for flag-primed group}
 \item{m_neutralprimed}{Mean conservatism score for neutral group}
 \item{sd_neutralprimed}{Standard deviation for those given the neutral group}
 \item{n_neutralprimed}{Sample size for neutral group}
}}
\source{
This is data is available online at \url{https://osf.io/wx7ck} from this study:
Klein, R. A., Ratliff, K. A., Vianello, M., Adams ., R. B., Bahnik, S., Bernstein, M. J., ... & Nosek, B. A. (2014).
Investigating Variation in Replicability. \emph{Social Psychology, 45}, 142-152. \url{http://doi.org/10.1027/1864-9335/a000178}

The original study exploring this effect is:
Carter, T. J., Ferguson, M. J., & Hassin, R. R. (2011). A Single Exposure to the American 
Flag Shifts Support Toward Republicanism up to 8 Months Later. \emph{Psychological Science, 22}, 
1011-1018. \url{http://doi.org/10.1177/0956797611414726}
}
\usage{
ma_flag_priming
}
\description{
An example dataset used in Chapter 9 of the book \emph{Introduction to the New
Statistics}.
}
\details{
To what extent can subtle cues influence political attitudes?  In this study, 
participants were asked to look at photos to estimate the time of day they 
were taken.  For some participants, 2 of the 4 photos were of the U.S.A. flag 
("flag prime").  For the rest, all 4 photos were of neutral objects ("no 
prime").

Next, for what seemed like an unrelated study, participants completed a short 
8-item questionairre about political beliefs (each rated on a scale from 1-7, 
where higher numbers represent more conservative attitudes).

The research question is: to what extent does flag exposure alter political 
attitudes?

This is summary data from a number of different labs which attempted to
replicate this project.  The data is for U.S. labs only, as priming wih the
U.S. flag may not have the same effect in other countries:
}
\references{
Cumming, G., & Calin-Jageman, R. (2017). \emph{Introduction to the
New Statistics}. New York; Routledge.
}
\keyword{datasets}

