% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_files.R
\docType{data}
\name{math_gender_iat_ma}
\alias{math_gender_iat_ma}
\title{Meta-Analysis Example - Replication of gender difference in implicit
attitudes about math}
\format{A data frame with 30 rows and 9 variables: 
\describe{ 
 \item{location}{Name of the lab}
 \item{m_male}{Mean IAT score for males}
 \item{sd_male}{Standard deviation for males on IAT}
 \item{n_male}{Sample size for males}
 \item{m_female}{Mean IAT score for females}
 \item{sd_female}{Standard deviation for females on IAT}
 \item{n_female}{Sample size for females on IAT}
 \item{subset}{Factor indicating whether the study was conducted in the USA or not}
 \item{country}{Factor with 12 levels indicating the country where the study was conducted}
}}
\source{
This is data is available online at \url{https://osf.io/wx7ck} from this
study: 
Klein, R. A., Ratliff, K. A., Vianello, M., Adams ., R. B., Bahnik, S., Bernstein, M. J., ... & Nosek, B. A. (2014).
Investigating Variation in Replicability. \emph{Social Psychology, 45}, 142-152. \url{http://doi.org/10.1027/1864-9335/a000178}

The original study exploring this effect is: 
Nosek, B. a, Banaji, M. R., & Greenwald, A. G. (2002). 
Math = male, me = female, therefore math not = me. 
\emph{Journal of Personality and Social Psychology, 83}(1), 44-59. 
\url{http://doi.org/10.1037/0022-3514.83.1.44}
}
\usage{
math_gender_iat_ma
}
\description{
An example dataset used in Chapter 9 of the book \emph{Introduction to the New
Statistics}.
}
\details{
To what extent is gender related to implicit attitudes about bias?  To find
out, Nosek and colleagues as male and female students to complete an Implicit
Association Test (IAT) that measured how easily negative ideas could be
connected to art or to mathematics.  Higher scores indicate higher levels of
bias against mathematics.

Note that all participants collected who had to be excluded due to high error rates 
or slow responses were excluded prior to creating this summary.  Data were omitted from 
these sites:uva vcu wisc wku wl wpi
}
\references{
Cumming, G., & Calin-Jageman, R. (2017). \emph{Introduction to the
 New Statistics}. New York; Routledge.
}
\keyword{datasets}

