% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natsal.R
\docType{data}
\name{natsal}
\alias{natsal}
\title{Impact of Outliers - Sexual Partners}
\format{A data frame with 95 rows and 2 variables:
\describe{
 \item{sex}{Sex: female or male}
 \item{partners}{Number of sexual partners}
}}
\source{
National Centre for Social Research et al. (2005), \emph{National Survey of Sexual Attitudes and Lifestyles II,
2000-2001 [computer file]}. Colchester, Essex: UK Data Archive [distributor], August 2005. SN: 5223.
}
\usage{
natsal
}
\description{
An example of independent groups data with outliers used in Chapters 16 of the
book \emph{Introduction to the New Statistics}.
}
\details{
This is a subset of data from a British survey of sexual behaviours and
attitudes that can be used to illustrate the impact of outliers on estimates
of central tendency. The participants in this subset were all 21 years of age.
The independent variable is sex and the dependent variable is the lifetime
number of sexual partners.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

