% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_moral.R
\docType{data}
\name{organic_moral}
\alias{organic_moral}
\title{One Way Design - Organic Foods and Moral Judgements}
\format{A data frame with 106 rows and 2 variables:
\describe{
 \item{food}{Three levels - \emph{comfort}, \emph{control}, or \emph{organic}}
 \item{judge}{Moral Judgement rating}
}}
\source{
Moery, E. & Calin-Jageman, R.J. (In Revision).  Direct and conceptual replications of Eskine (2013):
Organic food exposure has little to no effect on moral judgements.  Under consideration at \emph{Social Psychological and Personality Science}.

The data is available at \url{https://osf.io/atkn7}

The original study investigating this phenomenon was:

Eskine, K. J. (2013). Wholesome Foods and Wholesome Morals?: Organic Foods Reduce Prosocial Behavior and Harshen Moral Judgments.
\emph{Social Psychological and Personality Science, 4}, 251-254. \url{http://doi.org/10.1177/1948550612447114}
}
\usage{
organic_moral
}
\description{
An example dataset used in Chapter 13 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
Do organic foods make us morally smug?  To what extent?  To find out, Eskine (2013) asked
participants to rate images of organic food, neutral (control) food, or comfort food.
Next, all participants completed a moral judgements scale in which they read different moral
scenarios and judged how wrong they were (scale of 1-7).

The original study examining this phenomenon was Eskine (2013).  The primary research questions were:
(i) Does judgmentalness increase after organic food exposure compared to control food, and (ii)
Does judgmentalness decrease after comfort food exposure compared to control food?

Moery & Calin-Jageman (under review) conducted a close replication of the study.
The same research questions were asked in the close replication.  The data here comes from the
Moery and Calin-Jagerman replication study.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

