% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_laptop.R
\docType{data}
\name{pen_laptop1}
\alias{pen_laptop1}
\title{Two Independent Groups  - Pen vs Laptop Study 1}
\format{A data frame with 65 rows and 2 variables:
\describe{
 \item{Group}{Independent variable: Laptop or Pen}
 \item{Transcription}{Dependent variable: Transcription score}
}}
\source{
Mueller, P. A., & Oppenheimer, D. M. (2014). The pen is
 mightier than the keyboard: Advantages of longhand over laptop note taking.
 \emph{Psychological Science, 25}, 1159-1168. doi:10.1177/0956797614524581
}
\usage{
pen_laptop1
}
\description{
An example of data for the two independent groups design used in Chapters 2
and 7 of the book \emph{Introduction to the New Statistics}.
}
\details{
Mueller and Oppenheimer (2014) studied whether students learned more after
taking notes by hand rather than using a computer. They conducted a study in
which students were randomly assigned to take notes using a pen or by typing
on a laptop. The study found that students learned more after taking notes in
longhand than after typing notes. The researchers wondered whether writing
encouraged expression of concepts in the student's own words, whereas typing
encouraged relatively mindless transcription that led to relatively poor
learning. To investigate, the researchers devised a transcription score, the
percentage of notes that was verbatim transcription from the lecture.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{pen_laptop2}}
}
\keyword{datasets}

