% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_laptop.R
\docType{data}
\name{pen_laptop2}
\alias{pen_laptop2}
\title{Two Independent Groups  - Pen vs Laptop Study 2}
\format{A data frame with 151 rows and 2 variables:
\describe{
 \item{group}{Independent variable: \emph{Laptop} or \emph{Pen}}
 \item{transcription}{Dependent variable: Transcription score}
}}
\source{
Mueller, P. A., & Oppenheimer, D. M. (2014). The pen is
 mightier than the keyboard: Advantages of longhand over laptop note taking.
 \emph{Psychological Science, 25}, 1159-1168. \url{http://doi.org/10.1177/0956797614524581}
}
\usage{
pen_laptop2
}
\description{
An example of data for the two independent groups design used in Chapter 7 of
the book \emph{Introduction to the New Statistics}.
}
\details{
This data is for a second study from Mueller and Oppenheimer (2014). See
\code{\link{pen_laptop1}} for details.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{pen_laptop1}}
}
\keyword{datasets}
