% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PenLaptop1}
\alias{PenLaptop1}
\title{PenLaptop1 data.}
\format{A data frame with 65 rows and 2 variables:
\describe{
 \item{Group}{A factor with two levels - Male and Female}
 \item{Transcription}{Add description of DV here, and units}
}}
\usage{
PenLaptop1
}
\description{
Example dataset from Introduction to the New Statistics
}
\keyword{datasets}

