% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattan.R
\docType{data}
\name{rattan}
\alias{rattan}
\title{One Way Design - Rattan}
\format{A data frame with 51 rows and 2 variables: \describe{ 
 \item{group}{Factor with three levels - \emph{challng}, \emph{comfort}, \emph{control}} 
 \item{motivation}{Student motivation score (dependent variable)} }}
\source{
You can find the materials and data on the Open Science Framework: 
 \url{https://osf.io/xtkve}
}
\usage{
rattan
}
\description{
An example dataset used in Chapter 14 of the book \emph{Introduction to the 
New Statistics}.
}
\details{
How do you think you would react to feedback that gave encouragement and 
reassurance, or, instead, encouragement and challenge? Carol Dweck and her 
colleagues have investigated many such questions about how people respond to 
different types of feedback. This data set comes from Dweck's research group.
Rattan et al. (2012) asked their college student participants to imagine they
were undertaking a math course and had just received a low score (65%) on the
first test of the year. Participants were assigned randomly into three groups,
which received different feedback along with the low score. The Comfort group
received positive encouragement and also reassurance, the Challenge group
received positive encouragement and also challenge, and the Control group
received just the positive encouragement. Participants then responded to a
range of questions about how they felt about the course and their professor.
The dependent variable included here is ratings of their own motivation
towards mathematics after they had received the feedback.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017). \emph{Introduction to the
 New Statistics}. New York; Routledge.

Rattan, A., Good, C., & Dweck, C. S. (2012). "It's ok - Not everyone 
can be good at math": Instructors with an entity theory comfort (and demotivate) 
students. \emph{Journal of Experimental Social Psychology, 48}, 731-737.
}
\keyword{datasets}

