% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/religion_sharing.R
\docType{data}
\name{religion_sharing}
\alias{religion_sharing}
\title{One-Way Independent Groups Design Example - Religious Parenting}
\format{A data frame with 3 rows and 4 variables: 
\describe{ 
 \item{statistic}{Character string indicitaing whether the statistic is the sample size (n), 
                   Mean (m), or Standard Deviation(s)}
 \item{non_religious_parents}{Statistics for the non religious parents group}
 \item{christian_parents}{Statistics for the christian parents group}
 \item{muslim_parents}{Statistics for the muslim parents group}
   }}
\source{
The summary data comes from:

Decety, J., Cowell, J. M., Lee, K., Mahasneh, R., Malcolm-Smith, S., Selcuk, B., 
& Zhou, X. (2015). The Negative Association between Religiousness and Children's 
Altruism across the World. \emph{Current Biology}, 1-5. 
\url{doi:10.1016/j.cub.2015.09.056}
}
\usage{
religion_sharing
}
\description{
Summary data used in Chapter 14 of the book \emph{Introduction to the New
Statistics}.
}
\details{
To what extent does being raised in a religious famility relate to prosocial 
or altruistic behavior.  To investigate, Decety et al. (2015) collected data 
from a large, international sample (N = 1,170 with families recruited from the 
U.S., Canada, Jordan, Turkey, South Africa, and China.).  Parents reported 
their religious belief (Non-religious, Christian, Muslum, Buddhist, Jewish, 
Hindu, or other).  In addition, children played a game in which they were 
given 10 stickers but then asked if they would share some of these stickers 
with another child who hadn't been able to make it to the lab to play the game.  
The measure of altruism/sharing is the number of stickers given (out of 10).

Summary data from this study are provided in the data frame.

The researhcers asked two questions:
\itemize{
         \item Do children raised by Christian parents differ in sharing/altruism
          relative to children raised by Muslim parents?
          \item Do children raised by non-religious parents differ in sharing/altruism 
          relative to children raised by religious parents (Christian or Muslim).
}
}
\references{
Cumming, G., & Calin-Jageman, R. (2017). \emph{Introduction to the
New Statistics}. New York; Routledge.
}
\keyword{datasets}

