% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/religious_belief.R
\docType{data}
\name{religious_belief}
\alias{religious_belief}
\title{Describing Data Example - Study of Religous Belief}
\format{A data frame with 413 rows and 4 variables:
\describe{
 \item{response_id}{Respondent identifier}
 \item{condition}{Experimental condition: \emph{discus} if participants had viewed 3 images of a discus thrower sculpture, \emph{thinker} if participants had viewed 3 images of the thinker sculpture}
 \item{belief_in_gd}{Rating of how strongly participants believed in god, from 0 (certain that god does not exist) ot 100 (certain that God exists)}
 \item{age}{Age in years}
}}
\usage{
religious_belief
}
\description{
An example of data from a study of religious beliefs used in Chapter 3 of the book
\emph{Introduction to the New Statistics}.
}
\details{
This is a subset of data from an attempt to replicate Study 2 of Gervais & Norenzayan (2012).
The project details can be found at \url{https://osf.io/qc6rh/}
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.

Gervais, W. M., & Norenzayan, a. (2012).
Analytic Thinking Promotes Religious Disbelief. \emph{Science, 336(6080)},
493-496. doi:10.1126/science.1215647
}
\keyword{datasets}

