% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_beauty.R
\docType{data}
\name{sleep_beauty}
\alias{sleep_beauty}
\title{Correlated Data - Sleep and Beauty}
\format{A data frame with 70 rows and 2 variables:
\describe{
 \item{nightly_sleep_hours}{Typical hours of nightly sleep (0 to 12)}
 \item{rated_attractiveness}{Level of attractiveness rated on a scale from 1-10}
}}
\usage{
sleep_beauty
}
\description{
An example of correlated data used in Chapter 11 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
This is a fake dataset created for the purposes of teaching / illustration.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}
