% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_golf.R
\docType{data}
\name{super_golf}
\alias{super_golf}
\title{Two Independent Groups - Superstition and Golf}
\format{A data frame with 111 rows and 7 variables:
\describe{
 \item{id}{Respondent identifier}
 \item{condition}{\emph{Control} for normal ball, \emph{Lucky} for lucky ball}
 \item{gender}{Gender: Female or Male}
 \item{felt_lucky_1}{First item on manipulation check, rated agreement with statement "Before starting this task, I believed the golf ball assigned to me was lucky" on a scale from 1 to 5.}
 \item{felt_lucky_2}{Second item on manipulation check, rated agreement with statement "Now that I have completed this task, I believe that the golf ball assigned to me is lucky" on a scale from 1 to 5.}
 \item{felt_lucky_avg}{Average of the two \emph{felt_lucky} items}
 \item{golf_score}{Number of shots made out of 10.}
}}
\source{
Calin-Jageman, R. J., & Caldwell, T. L. (2014). Replication of the Superstition and Performance Study by Damisch, Stoberock, and Mussweiler (2010).
\emph{Social Psychology, 45}, 239-245. \url{http://doi:10.1027/1864-9335/a000190}

This dataset excludes a participant who asked for their data to be withdrawn and another participant who failed a memory manipulation check.

The complete dataset for this study is available online at \url{http://osf.io/fsadm}
}
\usage{
super_golf
}
\description{
An example of data from a study with a two independent groups design used in Chapter 7 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
To what extent can superstition influence performance?  To find out, participants were asked to complete a 10 shot golf task.
Before the game started, though, participants drew a ball from the bag.  Half the balls were plain old golf balls, and if the participant
drew one of these the experimenter simply said "This is the ball you will use."  These participants were considered a \emph{control}.
The other half of the balls were marked with lucky 4-leaf clovers, and if drawn the experimenter exclaimed "Wow! You get to use the lucky ball!".
These participants were considered \emph{lucky}.
After the golf task, participants completed a two item manipulation check: "Before starting this task, I believed that the golf
ball assigned to me was lucky" and "Now that I have completed this task, I believe that the golf ball assigned to me is lucky."
Responses were made on a Likert scale from 1 (strongly disagree) to 5 (strongly agree).
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

