% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomason.R
\docType{data}
\name{thomason1}
\alias{thomason1}
\title{Two Dependent Groups - Thomason Study 1}
\format{A data frame with 12 rows and 2 variables:
\describe{
 \item{pre}{Pre-treatment LSAT score}
 \item{post}{Post-treatment LSAT score}
}}
\source{
Thomason, N. R., Adajian, T., Barnett, A. E., Boucher, S.,
 van der Brugge, E., Campbell, J., Knorpp, W., Lempert, R., Lengbeyer, L.,
 Mandel, D. R., Rider, Y., van Gelder, T., & Wilkins, J. (2014).
 \emph{Critical thinking final report}. The University of Melbourne,
 N66001-12-C-2004.
}
\usage{
thomason1
}
\description{
An example of data for the repeated measures design used in Chapter 8 of the 
book \emph{Introduction to the New Statistics}.
}
\details{
Many colleges require a course in critical thinking, but such courses are not
particularly effective. Courses based on argument mapping, however, are more
effective. Argument mapping is a graphical technique for representing and
critiquing an argument. Thomason et al. (2014) reported 7 studies evaluating a
promising approach to teaching critical thinking that combines argument
mapping with a form of mastery learning. Studies were conducted in the U.S.,
Canada, and the U.K., and each study used a single group of students. Students
were tested on various well-established measures of critical thinking, both
before (the Pre-test) and after (the Post-test) training. Group sizes ranged
from 7 to 39. All the Thomason studies compared the two conditions, Pretest
and Posttest, within participants, and therefore used a paired design. The
first Thomason study, Thomason 1, used a group of N = 12 students, whose
critical thinking ability was assessed at Pretest and Posttest using the
Logical Reasoning section of the Law School Aptitude Test (LSAT).
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{thomason2}} \code{\link{thomason3}}
}
\keyword{datasets}

