% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomason.R
\docType{data}
\name{thomason3}
\alias{thomason3}
\title{Two Dependent Groups - Thomason Study 3}
\format{A data frame with 39 rows and 2 variables:
\describe{
 \item{pre}{Pre-treatment HCTA score}
 \item{post}{Post-treatment HCTA score}
}}
\source{
Thomason, N. R., Adajian, T., Barnett, A. E., Boucher, S.,
 van der Brugge, E., Campbell, J., Knorpp, W., Lempert, R., Lengbeyer, L.,
 Mandel, D. R., Rider, Y., van Gelder, T., & Wilkins, J. (2014).
 \emph{Critical thinking final report}. The University of Melbourne,
 N66001-12-C-2004.
}
\usage{
thomason3
}
\description{
An example of data for the repeated measures design used in Chapter 8 of the 
book \emph{Introduction to the New Statistics}.
}
\details{
This is data for the third study by Thomason et al investigating techniques to improve critical
thinking. This study used the HCTA measure of critical thinking, and a paired
design with N = 39. See \code{\link{thomason1}} for more details.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\seealso{
\code{\link{thomason1}} \code{\link{thomason2}}
}
\keyword{datasets}

