% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videogame_aggression.R
\docType{data}
\name{videogame_aggression}
\alias{videogame_aggression}
\title{Two Way Design - Video game Aggression}
\format{A data frame with 224 rows and 3 variables:
\describe{
 \item{violence}{Independent variable with two levels - \emph{nonviolent} or \emph{violent}}
 \item{difficulty}{Independent variable with two levels - \emph{easy} or \emph{hard}}
 \item{aggression}{Aggression score, rated from 0 seconds to 80 seconds in 10s increments}
}}
\source{
This is a simplified data set from Hilgard (2015).  You can find the materials and analysis plan for this study
on the Open Science Framework: \url{https://osf.io/cwenz}

Hilgard, J. (2015). \emph{Game violence, game difficulty, and 2D:4D digit ratio as predictors of aggressive behavior}.
University of Missouri-Columbia.
}
\usage{
videogame_aggression
}
\description{
An example dataset used in Chapter 14 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
Hilgard (2015) asked male participants to play a video game for 15 minutes.  The game was customized
so that it could vary in violence (shooting zombies or helping aliens) and difficulty (targets controlled
by tough AI or dumb AI).  After the game, players were provoked by being given an insulting evaluation by
a confederate.  Participants then got to decide how long the confederate should hold their hand in painfully
cold ice water (0-80s in 10s increments), and this was taken as a measure of aggressive behavior.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

