% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11_ex7.R
\docType{data}
\name{ch11_ex7}
\alias{ch11_ex7}
\title{Correlation Example - Social Inclusion and Exclusion}
\format{A data frame with 145 rows and 5 variables:
\describe{
 \item{participant}{Respondent Identifier}
 \item{condition}{\emph{Excluded} or \emph{Included}}
 \item{rater_1}{Rater 1 rating}
 \item{rater_2}{Rater 2 rating}
 \item{temperature_rating}{Temperature rating}
}}
\source{
Full details on this study are available at: \url{https://osf.io/95gvy}
}
\usage{
ch11_ex7
}
\description{
A dataset used in Chapter 11 of the book 
\emph{Introduction to the New Statistics}.
}
\details{
This data is from Cavenaugh & Calin-Jageman (in prep), a registered replication of a study by
Zhong & Leonardelli (2008) that found feeling socially excluded makes one feel physically cold.

Participants wrote an essay about being socially included or excluded.  They then reported what they
perceived to be the room temperature, ostensibly for a maintenance issue.  Temperatures were
reported in Fahrenheit but are reported here in Celsius.

As a manipulation check, two raters looked at the essay and it on a scale from -5
(completely excluded) to 5 (completely included).  Ratings were made blind to condition.
}
\references{
Cumming, G., & Calin-Jageman, R. (2017).
 \emph{Introduction to the New Statistics}. New York; Routledge.
}
\keyword{datasets}

