% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_fun}
\alias{demo_fun}
\title{Demo function title here}
\usage{
demo_fun(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A second numeric vector.}
}
\value{
Return demo value here
}
\description{
\code{demo_fun} The second paragraph is the description.
}
\section{MyTestCustomSection}{

 Demo testing.
}
\examples{
demo_fun(1, 10)
demo_fun(1000, 3)

\dontrun{
demo_fun("a", 3)
}
}

